sbt_build
Partial work towards building an ONI ML jar using SBT for dependency management and build.
