+++ build.sbt
@@ -1,0 +1,10 @@
+name := "ONI ML"
+
+version := "1.1"
+
+scalaVersion := "2.10.5"
+
+libraryDependencies += "org.apache.spark" %% "spark-core" % "1.3.0"
+
+
+
